redagavimo ir salinimo funkcionalumas

+++ fridgeshare-api/FridgeShare/Services/Storages/StorageService.cs
@@ -58,11 +58,17 @@ public class StorageService : IStorageService
 
     public async Task<ErrorOr<Deleted>> DeleteStorage(Guid id)
     {
-        var storage = await _dbContext.Storages.FindAsync(id);
+        var storage = await _dbContext.Storages
+            .Include(s => s.Products).FirstOrDefaultAsync(s => s.Id == id);
 
         if (storage is null)
         {
             return Errors.Storage.NotFound;
+        }
+
+        if(storage.Products != null && storage.Products.Count > 0)
+        {
+            return Errors.Storage.HasProducts;
         }
 
         _dbContext.Storages.Remove(storage);
